requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
boto3==1.35.17
botocore==1.35.17
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Collectfasta==3.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dj-database-url==2.2.0
Django==5.1.1
django-celery-results==2.5.1
django-cleanup==8.1.0
django-redis==5.4.0
django-storages==1.14.4
django-upgrade==1.21.0
flower==2.0.1
gunicorn==23.0.0
h11==0.14.0
humanize==4.10.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
jmespath==1.0.1
kombu==5.4.1
mailjet-rest==1.3.4
msgpack==1.1.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
sentry-sdk==2.14.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tokenize-rt==6.0.0
tomli==2.0.1
tornado==6.4.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==7.0.3

# Testing dependencies
pytest==9.0.2
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==1.3.0
coverage==7.6.9
factory-boy==3.3.1
faker==33.1.0

# Message Queue and Event Streaming
pika==1.3.2
kafka-python==2.0.2
